requirements.txt
apache-beam==2.10.0
avro==1.8.2
cachetools==3.1.0
certifi==2018.11.29
chardet==3.0.4
crcmod==1.7
dill==0.2.9
docopt==0.6.2
enum34==1.1.6
fastavro==0.21.18
fasteners==0.14.1
funcsigs==1.0.2
future==0.17.1
futures==3.2.0
google-api-core==1.8.0
google-api-python-client==1.7.8
google-apitools==0.5.24
google-auth==1.6.3
google-auth-httplib2==0.0.3
google-cloud-bigquery==1.6.1
google-cloud-core==0.29.1
google-cloud-pubsub==0.39.0
google-resumable-media==0.3.2
googleapis-common-protos==1.5.8
googledatastore==7.0.2
grpc-google-iam-v1==0.11.4
grpcio==1.18.0
hdfs==2.2.2
httplib2==0.11.3
idna==2.8
mock==2.0.0
monotonic==1.5
numpy==1.16.1
oauth2client==3.0.0
pbr==5.1.2
proto-google-cloud-datastore-v1==0.90.4
protobuf==3.6.1
pyarrow==0.11.1
pyasn1==0.4.5
pyasn1-modules==0.2.4
pydot==1.2.4
pyparsing==2.3.1
pytz==2018.9
PyVCF==0.6.8
PyYAML==3.13
requests==2.21.0
rsa==4.0
six==1.12.0
typing==3.6.6
uritemplate==3.0.0
urllib3==1.24.1
